Bug fixes
- added pragma pack directives so GCC doesn't re-align the members with a struct
- this fixes problems seen sending and receiving messages between apps compiled on
  machines with different compiler versions, OSs or architectures
- changed the fields storing the length of variable attributes to be long longs
- these are 64bits cross platform regardless of architecture or compiler
- this also increases the maximum size of a variable length attribute on 32bit machines

+++ lib/templates/header_entries/raw_header_entry.h
@@ -1,1 +1,1 @@
-int <%= @name %>_length;
+long long <%= @name %>_length;
